Refactorizes folders and packages
